Use buildkit image mirror and update docs
This commit uses the mirrored snapshot of the upstream buildkit image
(until there are upstream tags available we'll maintain a periodically
updated mirror).

It also updates the docs adding some info on the `of-buildkit` container
and the now slightly repurposes `of-builder` container (it no longer
needs to be run priviliged).

Signed-off-by: John McCabe <john@johnmccabe.net>