Fix issues / naming
Wrong signature was used in a previous PR for the method signature
func Handle(req []byte) []byte - this should return string and
was printing bytes in the logs.

Also improved the naming from "sss" + "ss" to "ss" and
"userSecret". "bs" becomes encodedBytes amongst other changes.

Tested on Kubernetes on AWS.

Signed-off-by: Alex Ellis (VMware) <alexellis2@gmail.com>