Add config and timeout with context
Config moves parsing into separate struct to keep the main body
clearner and to enable re-use. Tested with unit tests and with a
local instance of netcat.

Added context.Deadline for upstream proxy calls to give a HTTP
code when upstream call is cancelled. Tested with netcat and a
low timeout value of around 50ms.

Signed-off-by: Alex Ellis (VMware) <alexellis2@gmail.com>