Report status to github Checks API with logs
github-status function is already coupled with github. Github offer a much richer API than the Status API. Checks API allows to provide rich information about a perticular status directly in the github UI, making it much more easier for developers to look through logs and pull requests for functions.

Now statuses are reported with additional details - the build logs.

The change is contained within the github-status function and is abstracted away from the rest of the openfaas-cloud.

Signed-off-by: Kiril Vuchkov <kirilvuchkov@gmail.com>