Added option to clone private respositories
Closes #7

We have github's auth token for displaying statuses so we could
use the same approach and generate auth token just for cloning
private repositories.

I moved GetPrivateKeyPath function from github-status, and
placed it inside sdk/auth because we can reuse itside git-tar
to clone private repository. I have created tests for
that function also.

I like to create pure functions so I created an abstraction
of authToken, which has method GetToken inside. In the future
if we will have auth token already, we could reuse it,
but probably we'll have to add mechanism to check if token
maybe should be generated again (it's live ended).
Because of that abstraction, I could create a tokener interface
which is used inside getRepositoryURL function. It helps with
creation of unit tests.

getRepositoryURL function will return the same clone URL
for repository if it's public, but if it's private,
it'll get a token first, then it'll put installationID:authToken
to the url so we can clone private repository
(ie. https://installationID:authToken@github.com/organization/repository)

I added .idea directory to the .gitignore because I'm using
jetbrains' IDEs also.

Signed-off-by: Bart Smykla (VMware) <bartek@smykla.com>