Add the new version of OpenFaaS Cloud Dashboard
This change adds all the assets for the React SPA.

* The React SPA is created with `create-react-app` and is bundled with
the `react-scripts` as a common react app would do.
* The assets (index.html, *.js, *.css) are served from the
`of-cloud-dashboard` function using the `node8-express` template
* Since the base href and the user's endpoints change depending on the
usage of `pretty_url`, these data are injected from the `node8-express`
function to the SPA via global variables inside the `window` object
(i.e. `window.BASE_HREF`)

The design and implementations are almost exactly the same as the
current dashboard (jquery version).

The docs are also updated to reflect the new installation instructions.

Signed-off-by: Ken Fukuyama <kenfdev@gmail.com>