Add initial feature for auth proxy
- The router component now delegates auth to (auth) microservice
for each request. Protected routes need a cookie to proceed or are
asked to redirect to GitHub to obtain a cookie as part of the
authorization flow.

- Work remaining the auth service must sign the cookie ising a
private key and then validate it on each HTTP request using its
public key.

A GitHub app must be created in order to use the auth delegation,
this must be configured via environmental variables but will move
to secrets.

Signed-off-by: Alex Ellis (VMware) <alexellis2@gmail.com>