Add handling of zero-length function names to router
If a OFC user strips the url to the base FQDN then an error message is returned that suggests the system has tried to route through to a function named 'user-'.  This change handles this scenario within the router so that a 404 is returned.

Additional small change was made to the way leading slashes in the requestURI are handled in the router, removing the assumption that only a single slash would occur.

Signed-off-by: Richard Gee <richard@technologee.co.uk>