Fix dashboard link
The stack-deploy link was trying to point to gateway_pretty_url,
edited to a non-existing endpoint. It's now fixed to point to the
dashboard URL

Used methods implementation from @martindekov's pull request for
gitlab status link. If they are moved to sdk, this will allow
refactoring of the logic in github-status

Signed-off-by: Ivana Yovcheva <iyovcheva@vmware.com>