Fix regression in github-event
The github-event function was validating the sender instead
of the repo owner. There is a goal to restrict the repo owner or
org owner to only those in the CUSTOMERS file.

Added unit tests and tested e2e with:

- negative test showing "buildsimple" was rejected as not in
the customers
- positive test showing "alexellis" in file was allowed to build
due to being in customers
- positive test showing "openfaas" (pushed by "alexellis" worked
due to the "openfaas" org being in the customers file

New version 0.8.0 of github-event has been published.

Signed-off-by: Alex Ellis <alexellis2@gmail.com>