Add "ofc" prefix to all pipeline functions
By adding this prefix to the deployments and by then restricting
access in edge-auth, functions such as pipeline-log can no longer
be accessed via URLs such as pipeline.example.com/log

Users will be advised to update. This does not affect functions
which already implement a payload secret or HMAC to verify the
invoker.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>