Make Github Status namespace checks by APP ID
The Github Status function was not namespacing the check-runs. It was
using the same ting for each installed app. This has been changed to
namespace by app_id. This means that 2 + installations on the same
repository can now correctly create & report github check runs.

This has been tested by installing 2x OCF installations, then running a
pipeline. This failed of half of the functions as they didn't get the
correct check for that app.

Then I added the new function to each OFC installation, and re-ran a
pipeline. Every check run was created and reported back correctly.

Signed-off-by: Alistair Hey <alistair@heyal.co.uk>