Don't error if ECR Repository exists already
We were returning an error if the ECR repo already exists. This shouldnt
give an error because it's not a failure. If the repo doesn't exist then
we create it, if it exists then that's fine.

Tested by deploying an ECR backed OFC installation, first time round it
creates the new repo, then after that it's not throwing an error, only
saying it already exists.

Signed-off-by: Alistair Hey <alistair@heyal.co.uk>