Add Cookie root domain env variable to Dashboard
Somewhere along the line the cookie_root_domain was missed from the
dashboard config. This only needs to be set when using OAuth, to remove
a users token from their cookie. This was not set and therefore caused
us to try and set the cookie domain to "undefined"

This has been tested by adding the env to the Dashboard deployment and
hitting logout on an OAuth enabled installation. This now sets the
correct cookie domain and removes the existing token

Signed-off-by: Alistair Hey <alistair@heyal.co.uk>