Part one of converting OFC to go modules
This will initialise a module in each function/folder. This needs to be
merged before modules can start to manage dependancies because otherwise
you get a duplicate definition error as each function that imports
anything from the ofc repo (from github in its gopkg) will pull the
whole repo in, not just the module it wants (the SDK for example) so we
need a go mod in each folder to avoid it existing twice in the
dependency tree for the function.

Once this is in, each module can be converted to go modules and remove
the use of dep

Signed-off-by: Alistair Hey <alistair@heyal.co.uk>